move custom indexing classes to oac dir
