+++ arclight/lib/oac/normalized_component_title.rb
